Extracted playbook business logic to playbook service; altered tests for consistency with that change. Added tests to track down a bug. Fixed bugs.
