+++ src/components/Timeline/Event.js
@@ -1,11 +1,13 @@
 import React from 'react'
+import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
 import moment from 'moment'
 import { Card, CardHeader, CardText } from 'material-ui/Card'
 import BootstrapPlaybook from './Playbook/BootstrapPlaybook'
 import Playbook from './Playbook/Playbook'
+import { LoadTextFromEvent } from '../../services/playbookService'
 
-const Event = ({ text, time, backgroundColor, incidentId, ticketId, eventTypeId, eventId, eventActions, eventTypeActions }) => {
+export const Event = ({ text, time, backgroundColor, incidentId, ticketId, eventType, eventTypeId, eventId, eventActions, eventTypeActions }) => {
     return (
     <div>
         <BootstrapPlaybook
@@ -45,6 +47,24 @@ Event.propTypes = {
     time: PropTypes.instanceOf(moment),
     backgroundColor: PropTypes.string,
     ticketId: PropTypes.string
+}
+
+export const mapStateToEventProps = (state, ownProps) => {
+    const event = ownProps.event
+    const eventType = state.eventTypes.records[ownProps.event.eventTypeId]
+    const ticket = state.tickets.map[ownProps.ticketId]
+    const engagement = state.engagements.list.find(engagement => engagement.id === ownProps.engagementId)
+    return {
+        ...ownProps,
+        eventType,
+        ticket,
+        engagement,
+        eventId: event.id,
+        time: moment(event.occurred),
+        dismissed: event.dismissed,
+        backgroundColor: event.backgroundColor,
+        text: LoadTextFromEvent(event, eventType, ticket, engagement)
+    }
 }
 
 export default Event
